Text Summarization and translation/requirements.txt
gradio
deeptranslator
newspaper3k
transfomers
